First pass on adding test to the cli
