+++ prime-router/docker-compose.yml
@@ -35,12 +35,12 @@ services:
       - TX_PHD__ELR_TEST__PASS=pass
       - TX_PHD__ELR_DEBUG__USER=foo
       - TX_PHD__ELR_DEBUG__PASS=pass
-      - PRIME__HL7_TEST__USER=foo
-      - PRIME__HL7_TEST__PASS=pass
-      - PRIME__HL7_BATCH_TEST__USER=foo
-      - PRIME__HL7_BATCH_TEST__PASS=pass
-      - PRIME__CSV_TEST__USER=foo
-      - PRIME__CSV_TEST__PASS=pass
+      - PRIME__HL7__USER=foo
+      - PRIME__HL7__PASS=pass
+      - PRIME__HL7_BATCH__USER=foo
+      - PRIME__HL7_BATCH__PASS=pass
+      - PRIME__CSV__USER=foo
+      - PRIME__CSV__PASS=pass
       - SENDGRID_API_KEY=SnBfxpPsQkmOh0SfU3X_ng
       - OKTA_baseUrl=hhs-prime.okta.com
       - OKTA_clientId=0oa6fm8j4G1xfrthd4h6
