+++ prime-router/metadata/organizations-local.yml
@@ -186,10 +186,14 @@
           redox_source_name: "Prime Data Hub (Staging)"
         format: REDOX
 
+  # 'prime' is a test organization, designed to be safely usable across all our environments.
+  # It has four receivers, each named _exactly_ the same as the format of data.
+  # Each also has a fake "county", also named _exactly_ the same as the format of the data.
+  # This is to aid in readability.
   - name: prime
     description: Prime Test Receiver
     services:
-      - name: csv-test
+      - name: CSV
         topic: covid-19
         schema: fl/fl-covid-19
         jurisdictionalFilter: [ "matches(ordering_facility_state, PM)", "matches(ordering_facility_county, CSV)" ]
@@ -203,7 +207,7 @@
             host: sftp
             port: 22
             filePath: ./upload
-      - name: hl7-test
+      - name: HL7
         topic: covid-19
         schema: tx/tx-covid-19-hl7
         jurisdictionalFilter: [ "matches(ordering_facility_state, PM)", "matches(ordering_facility_county, HL7)" ]
@@ -218,7 +222,7 @@
             host: sftp
             port: 22
             filePath: ./upload
-      - name: hl7-batch-test
+      - name: HL7_BATCH
         topic: covid-19
         schema: az/az-covid-19-hl7
         jurisdictionalFilter: [ "matches(ordering_facility_state, PM)", "matches(ordering_facility_county, HL7_BATCH)" ]
@@ -233,7 +237,7 @@
             host: sftp
             port: 22
             filePath: ./upload
-      - name: redox-test
+      - name: REDOX
         topic: covid-19
         schema: pa/pa-covid-19-redox
         jurisdictionalFilter: [ "matches(ordering_facility_state, PM)", "matches(ordering_facility_county, REDOX)" ]
