+++ prime-router/metadata/organizations.yml
@@ -183,3 +183,76 @@
         jurisdictionalFilter: [ "matches(observation, covid-19:.*)", "matches(state, .*)"]
         transforms: {deidentify: true}
         format: CSV
+
+  # 'prime' is a test organization, designed to be safely usable across all our environments.
+  # It has four receivers, each named _exactly_ the same as the format of data.
+  # Each also has a fake "county", also named _exactly_ the same as the format of the data.
+  # This is to aid in readability.
+  - name: prime
+    description: Prime Test Receiver
+    services:
+      - name: CSV
+        topic: covid-19
+        schema: fl/fl-covid-19
+        jurisdictionalFilter: [ "matches(ordering_facility_state, PM)", "matches(ordering_facility_county, CSV)" ]
+        batch:
+          operation: MERGE
+          numberPerDay: 1440 # Every minute
+          initialBatch: 00:00
+        format: CSV
+        transports:
+          - type: SFTP
+            host: sftp
+            port: 22
+            filePath: ./upload
+      - name: HL7
+        topic: covid-19
+        schema: tx/tx-covid-19-hl7
+        jurisdictionalFilter: [ "matches(ordering_facility_state, PM)", "matches(ordering_facility_county, HL7)" ]
+        batch:
+          operation: MERGE
+          numberPerDay: 1440 # Every minute
+          initialBatch: 00:00
+          timeZone: ARIZONA
+        format: HL7
+        transports:
+          - type: SFTP
+            host: sftp
+            port: 22
+            filePath: ./upload
+      - name: HL7_BATCH
+        topic: covid-19
+        schema: az/az-covid-19-hl7
+        jurisdictionalFilter: [ "matches(ordering_facility_state, PM)", "matches(ordering_facility_county, HL7_BATCH)" ]
+        batch:
+          operation: MERGE
+          numberPerDay: 1440 # Every minute
+          initialBatch: 00:00
+          timeZone: ARIZONA
+        format: HL7_BATCH
+        transports:
+          - type: SFTP
+            host: sftp
+            port: 22
+            filePath: ./upload
+      - name: REDOX
+        topic: covid-19
+        schema: pa/pa-covid-19-redox
+        jurisdictionalFilter: [ "matches(ordering_facility_state, PM)", "matches(ordering_facility_county, REDOX)" ]
+        batch:
+          operation: MERGE
+          numberPerDay: 1440 # Every minute
+          initialBatch: 00:00
+          timeZone: ARIZONA
+        defaults:
+          processing_mode_code: T
+          redox_destination_id: 62d62d52-3771-4151-aff4-4a3d420a8b7a
+          redox_destination_name: "Prime Local Redox Destination"
+          redox_source_id: d89d4057-930f-4c5b-bb39-8cddb326e928
+          redox_source_name: "Prime Hub (Staging)"
+        format: REDOX
+        transports:
+          - type: REDOX
+            apiKey: some_key
+            baseUrl: http://redox:1080
+
